requirements.txt
fastapi
uvicorn
semantic-kernel
huggingface-hub
dotenv
firebase-admin
pydantic
chainlit
pytest
pytest-asyncio
azure-cosmos
email-validator